chore: 폴더 구조 수정, Section 별도의 컴포넌트로 분리, 변경된 설정에 따라 lint 완료 	modified:   index.js 	modified:   metro.config.js 	renamed:    App.tsx -> src/App.tsx 	renamed:    __tests__/App.test.tsx -> src/__tests__/App.test.tsx 	new file:   "src/assets/fonts/# \355\217\260\355\212\270" 	new file:   "src/assets/images/# \354\235\264\353\257\270\354\247\200" 	new file:   "src/components/# \354\236\254\354\202\254\354\232\251 UI" 	new file:   src/components/Section.tsx 	new file:   "src/hooks/# \354\273\244\354\212\244\355\205\200 \355\233\205" 	new file:   "src/navigation/# \353\235\274\354\232\260\355\214\205" 	new file:   "src/screens/# \355\216\230\354\235\264\354\247\200" 	new file:   src/services/# API 	new file:   src/store/# Zustand
