+++ index.js
@@ -2,8 +2,9 @@
  * @format
  */
 
-import {AppRegistry} from 'react-native';
-import App from './App';
-import {name as appName} from './app.json';
+import { AppRegistry } from 'react-native';
+
+import { name as appName } from './app.json';
+import App from './src/App';
 
 AppRegistry.registerComponent(appName, () => App);
