+++ src/__tests__/App.test.tsx
@@ -4,6 +4,7 @@
 
 import React from 'react';
 import ReactTestRenderer from 'react-test-renderer';
+
 import App from '../App';
 
 test('renders correctly', async () => {
